streamlit app/requirements.txt
torch~=2.7.1
torchvision~=0.22.1
numpy~=2.3.2
streamlit~=1.47.1
pillow~=11.3.0